+++ dashboard/views/shipping-zone-rates/lang-form.php
@@ -12,10 +12,34 @@ $langFrm->setFormTagAttribute('onsubmit', 'setupLangRate(this); return(false);')
 
 $langFld = $langFrm->getField('lang_id');
 $langFld->setfieldTagAttribute('onChange', "editRateLangForm(" . $zoneId . ", " . $rateId . ", this.value);");
-
-$fld = $langFrm->getField('auto_update_other_langs_data');
-if (null != $fld) {
-    HtmlHelper::configureSwitchForCheckbox($fld);
+$translatorSubscriptionKey = FatApp::getConfig('CONF_TRANSLATOR_SUBSCRIPTION_KEY', FatUtility::VAR_STRING, '');
+if (!empty($translatorSubscriptionKey) && $langId != CommonHelper::getDefaultFormLangId()) {
+    $langFld->developerTags['fldWidthValues'] = ['d-flex', '', '', ''];
+    $langFld->htmlAfterField = '<a href="javascript:void(0);" onclick="editRateLangForm(' . $zoneId . ', ' . $rateId . ', ' . $langId . ', 1)" class="btn" title="' .  Labels::getLabel('BTN_AUTOFILL_LANGUAGE_DATA', $siteLangId) . '">
+                                <svg class="svg" width="18" height="18">
+                                    <use xlink:href="' . CONF_WEBROOT_URL . 'images/retina/sprite.yokart.svg#icon-translate">
+                                    </use>
+                                </svg>
+                            </a>';
 }
+?>
 
-echo $langFrm->getFormHtml();
+<div class="modal-header">
+    <h5 class="modal-title"><?php echo Labels::getLabel('LBL_SHIPPING_RATES_SETUP', $siteLangId); ?></h5>
+</div>
+<div class="modal-body form-edit">
+    <?php if(count($languages)){ ?>
+        <div class="form-edit-head">
+            <nav class="nav nav-tabs navTabsJs">
+                <a class="nav-link" href="javascript:void(0)" onclick="addEditShipRates(<?php echo $zoneId ?>, <?php echo $rateId ?>);"><?php echo Labels::getLabel('LBL_General', $siteLangId); ?></a>
+                <a class="nav-link active" href="javascript:void(0);" <?php echo (0 < $rateId) ? "onclick='editRateLangForm(" . $zoneId . "," . $rateId . ", " . array_key_first($languages) . ");'" : ""; ?>>
+                    <?php echo Labels::getLabel('LBL_LANGUAGE_DATA', $siteLangId); ?>
+                </a>
+            </nav>
+        </div>
+    <?php } ?>
+    <div class="form-edit-body loaderContainerJs" id="selectedTabContentJs">
+        <?php echo $langFrm->getFormHtml(); ?>
+    </div>
+    <?php require_once(CONF_THEME_PATH . '_partial/listing/form-edit-foot.php'); ?>
+</div>