+++ CHANGELOG.md
@@ -1027,6 +1027,7 @@ TV-9.4.0.20220406
     - #065376 - The image attributes added by the admin are not visible on the front end.
     - #065394 - The default meta tags are missing on the default meta tag listing page.
     - #065431 - The buyer registered with a phone number redirected to the email configuration page when trying to access the seller dashboard.
+    - #065485 - On shipping rate setup at seller end rate name should come insted of identifier 
 ## UAT points:
 
     - #Admin dashboard sales stats > set tabs priority, Move Affiliate tab to the last